@@ -17,14 +17,10 @@ return [
             'apiUrl' => 'https://api.zoom.us/v2/',
             'defaulValuesMeeting' => [
                 'topic' => 'Meeting Shedule with User',
-                'duration' => 30
-            ]
-        ]
+                'duration' => 30,
+            ],
+        ],
 
+    ],
 
-    ]
-
-    
-    
-    
 ];
