@@ -1,16 +1,14 @@
 <?php
 
-use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
 
-class CreateImeetingMeetingsTable extends Migration
+return new class extends Migration
 {
     /**
      * Run the migrations.
-     *
-     * @return void
      */
-    public function up()
+    public function up(): void
     {
         Schema::create('imeeting__meetings', function (Blueprint $table) {
             $table->engine = 'InnoDB';
@@ -25,22 +23,19 @@ class CreateImeetingMeetingsTable extends Migration
 
             $table->integer('entity_id');
             $table->string('entity_type');
-            
+
             $table->text('options')->nullable();
 
             $table->timestamps();
             $table->auditStamps();
-
         });
     }
 
     /**
      * Reverse the migrations.
-     *
-     * @return void
      */
-    public function down()
+    public function down(): void
     {
         Schema::dropIfExists('imeeting__meetings');
     }
-}
+};
