@@ -2,25 +2,19 @@
 
 namespace Modules\Imeeting\Database\Seeders;
 
-use Illuminate\Database\Seeder;
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Seeder;
 
 class ImeetingDatabaseSeeder extends Seeder
 {
     /**
      * Run the database seeds.
-     *
-     * @return void
      */
-    public function run()
+    public function run(): void
     {
         Model::unguard();
 
         $this->call(ImeetingModuleTableSeeder::class);
         $this->call(ZoomProviderSeeder::class);
- 
     }
-
-   
-
 }
