@@ -8,11 +8,11 @@ class ProviderTranslation extends Model
 {
     public $timestamps = false;
 
-     protected $fillable = [
-      'title',
-      'description',
-      'provider_id',
-      'locale',
+    protected $fillable = [
+        'title',
+        'description',
+        'provider_id',
+        'locale',
     ];
 
     protected $table = 'imeeting__provider_translations';
