@@ -4,11 +4,10 @@ namespace Modules\Imeeting\Http\Controllers\Admin;
 
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
+use Modules\Core\Http\Controllers\Admin\AdminBaseController;
 use Modules\Imeeting\Entities\Imeeting;
 use Modules\Imeeting\Http\Requests\CreateImeetingRequest;
 use Modules\Imeeting\Http\Requests\UpdateImeetingRequest;
-use Modules\Imeeting\Repositories\MeetingRepository;
-use Modules\Core\Http\Controllers\Admin\AdminBaseController;
 
 class MeetingController extends AdminBaseController
 {
@@ -16,6 +15,7 @@ class MeetingController extends AdminBaseController
      * @var ImeetingRepository
      */
     private $imeeting;
+
     private $paymentMethod;
 
     public function __construct(ImeetingRepository $imeeting)
@@ -28,10 +28,8 @@ class MeetingController extends AdminBaseController
 
     /**
      * Display a listing of the resource.
-     *
-     * @return Response
      */
-    public function index()
+    public function index(): Response
     {
         //$imeetings = $this->imeeting->all();
 
@@ -40,21 +38,16 @@ class MeetingController extends AdminBaseController
 
     /**
      * Show the form for creating a new resource.
-     *
-     * @return Response
      */
-    public function create()
+    public function create(): Response
     {
         return view('imeeting::admin.imeetings.create');
     }
 
     /**
      * Store a newly created resource in storage.
-     *
-     * @param  CreateImeetingRequest $request
-     * @return Response
      */
-    public function store(CreateImeetingRequest $request)
+    public function store(CreateImeetingRequest $request): Response
     {
         $this->imeeting->create($request->all());
 
@@ -64,11 +57,8 @@ class MeetingController extends AdminBaseController
 
     /**
      * Show the form for editing the specified resource.
-     *
-     * @param  Imeeting $imeeting
-     * @return Response
      */
-    public function edit(Imeeting $imeeting)
+    public function edit(Imeeting $imeeting): Response
     {
         return view('imeeting::admin.imeetings.edit', compact('imeeting'));
     }
@@ -76,43 +66,34 @@ class MeetingController extends AdminBaseController
     /**
      * Update the specified resource in storage.
      *
-     * @param  Imeeting $imeeting
-     * @param  UpdateImeetingRequest $request
-     * @return Response
+     * @param  Imeeting  $imeeting
      */
-    public function update($id, UpdateImeetingRequest $request)
+    public function update($id, UpdateImeetingRequest $request): Response
     {
-
         //Find payment Method
         $paymentMethod = $this->paymentMethod->find($id);
-        
-        //Add status request
-        if($request->status=='on')
-            $request['status'] = "1";
-        else
-            $request['status'] = "0";
 
-        $this->imeeting->update($paymentMethod,$request->all());
+        //Add status request
+        if ($request->status == 'on') {
+            $request['status'] = '1';
+        } else {
+            $request['status'] = '0';
+        }
+
+        $this->imeeting->update($paymentMethod, $request->all());
 
         return redirect()->route('admin.icommerce.paymentmethod.index')
             ->withSuccess(trans('core::core.messages.resource updated', ['name' => trans('imeeting::imeetings.single')]));
-        
     }
 
     /**
      * Remove the specified resource from storage.
-     *
-     * @param  Imeeting $imeeting
-     * @return Response
      */
-    public function destroy(Imeeting $imeeting)
+    public function destroy(Imeeting $imeeting): Response
     {
         $this->imeeting->destroy($imeeting);
 
         return redirect()->route('admin.imeeting.imeeting.index')
             ->withSuccess(trans('core::core.messages.resource deleted', ['name' => trans('imeeting::imeetings.title.imeetings')]));
     }
-
-    
-
 }
