@@ -2,21 +2,19 @@
 
 use Illuminate\Routing\Router;
 
-$router->group(['prefix' => '/imeeting/v1'], function (Router $router) {
+Route::prefix('/imeeting/v1')->group(function (Router $router) {
+    $router->apiCrud([
+        'module' => 'imeeting',
+        'prefix' => 'meetings',
+        'controller' => 'MeetingApiController',
+        'middleware' => ['create' => [], 'index' => []], // Just Testing
+    ]);
 
     $router->apiCrud([
-      'module' => 'imeeting',
-      'prefix' => 'meetings',
-      'controller' => 'MeetingApiController',
-      'middleware' => ['create' => [],'index' => []] // Just Testing
+        'module' => 'imeeting',
+        'prefix' => 'providers',
+        'controller' => 'ProviderApiController',
+        'middleware' => [],
     ]);
-    
-    $router->apiCrud([
-      'module' => 'imeeting',
-      'prefix' => 'providers',
-      'controller' => 'ProviderApiController',
-      'middleware' => []
-    ]);
-// append
-
+    // append
 });
