@@ -4,14 +4,15 @@ namespace Modules\Imeeting\Providers;
 
 use Illuminate\Support\Arr;
 use Illuminate\Support\ServiceProvider;
-use Modules\Core\Traits\CanPublishConfiguration;
 use Modules\Core\Events\BuildingSidebar;
 use Modules\Core\Events\LoadingBackendTranslations;
+use Modules\Core\Traits\CanPublishConfiguration;
 use Modules\Imeeting\Events\Handlers\RegisterImeetingSidebar;
 
 class ImeetingServiceProvider extends ServiceProvider
 {
     use CanPublishConfiguration;
+
     /**
      * Indicates if loading of the provider is deferred.
      *
@@ -21,10 +22,8 @@ class ImeetingServiceProvider extends ServiceProvider
 
     /**
      * Register the service provider.
-     *
-     * @return void
      */
-    public function register()
+    public function register(): void
     {
         $this->registerBindings();
         $this->app['events']->listen(BuildingSidebar::class, RegisterImeetingSidebar::class);
@@ -32,31 +31,27 @@ class ImeetingServiceProvider extends ServiceProvider
         $this->app['events']->listen(LoadingBackendTranslations::class, function (LoadingBackendTranslations $event) {
             $event->load('imeetings', Arr::dot(trans('imeeting::imeetings')));
             // append translations
-
         });
     }
 
-    public function boot()
+    public function boot(): void
     {
-        
         $this->publishConfig('imeeting', 'config');
         $this->publishConfig('imeeting', 'crud-fields');
 
-        $this->mergeConfigFrom($this->getModuleConfigFilePath('imeeting', 'settings'), "asgard.imeeting.settings");
-        $this->mergeConfigFrom($this->getModuleConfigFilePath('imeeting', 'settings-fields'), "asgard.imeeting.settings-fields");
-        $this->mergeConfigFrom($this->getModuleConfigFilePath('imeeting', 'permissions'), "asgard.imeeting.permissions");
+        $this->mergeConfigFrom($this->getModuleConfigFilePath('imeeting', 'settings'), 'asgard.imeeting.settings');
+        $this->mergeConfigFrom($this->getModuleConfigFilePath('imeeting', 'settings-fields'), 'asgard.imeeting.settings-fields');
+        $this->mergeConfigFrom($this->getModuleConfigFilePath('imeeting', 'permissions'), 'asgard.imeeting.permissions');
 
         //$this->loadMigrationsFrom(__DIR__ . '/../Database/Migrations');
     }
 
     /**
      * Get the services provided by the provider.
-     *
-     * @return array
      */
-    public function provides()
+    public function provides(): array
     {
-        return array();
+        return [];
     }
 
     private function registerBindings()
@@ -85,8 +80,6 @@ class ImeetingServiceProvider extends ServiceProvider
                 return new \Modules\Imeeting\Repositories\Cache\CacheProviderDecorator($repository);
             }
         );
-// add bindings
-
-
+        // add bindings
     }
 }
