@@ -2,8 +2,8 @@
 
 namespace Modules\Imeeting\Repositories\Cache;
 
-use Modules\Imeeting\Repositories\ProviderRepository;
 use Modules\Core\Icrud\Repositories\Cache\BaseCacheCrudDecorator;
+use Modules\Imeeting\Repositories\ProviderRepository;
 
 class CacheProviderDecorator extends BaseCacheCrudDecorator implements ProviderRepository
 {
