@@ -27,5 +27,5 @@ return [
     'title' => 'Title',
     'description' => 'Description',
     'status' => 'Status',
-    
+
 ];
