@@ -22,5 +22,5 @@ return [
     'messages' => [
     ],
     'validation' => [
-    ]
+    ],
 ];
