@@ -4,26 +4,15 @@ namespace Modules\Imeeting\Traits;
 
 trait MeetingableTransformer
 {
-
-
-	/**
-   	* Method to merge values with response
-   	*
-   	* @return array
-   	*/
-  
-	public function getDataMeetings()
-	{
-		
-		$data = [
-            'meetings' => $this->when($this->meetings, $this->meetings)
+    /**
+     * Method to merge values with response
+     */
+    public function getDataMeetings()
+    {
+        $data = [
+            'meetings' => $this->when($this->meetings, $this->meetings),
         ];
 
         return $data;
-        
-	}
-
-  	
-
-
+    }
 }
